持续填坑中（3/5）

+++ server/src/main/java/nju/quadra/hms/bl/WebsitePromotionBL.java
@@ -5,7 +5,9 @@ import com.google.gson.reflect.TypeToken;
 import nju.quadra.hms.blservice.WebsitePromotionBLService;
 import nju.quadra.hms.data.mysql.WebsitePromotionDataServiceImpl;
 import nju.quadra.hms.dataservice.WebsitePromotionDataService;
+import nju.quadra.hms.model.LoginSession;
 import nju.quadra.hms.model.ResultMessage;
+import nju.quadra.hms.model.UserType;
 import nju.quadra.hms.po.WebsitePromotionPO;
 import nju.quadra.hms.vo.WebsitePromotionVO;
 
@@ -16,14 +18,23 @@ import java.util.HashMap;
  * Created by RaUkonn on 2016/11/21.
  */
 public class WebsitePromotionBL implements WebsitePromotionBLService {
-    private final WebsitePromotionDataService websitePromotionDataService;
+    private final LoginSession session;
+    private final WebsitePromotionDataService websitePromotionDataService = new WebsitePromotionDataServiceImpl();
 
     public WebsitePromotionBL() {
-        websitePromotionDataService = new WebsitePromotionDataServiceImpl();
+        session = null;
+    }
+
+    public WebsitePromotionBL(LoginSession session) {
+        this.session = session;
     }
 
     @Override
     public ArrayList<WebsitePromotionVO> get() {
+        if (session != null && !session.userType.equals(UserType.WEBSITE_MARKETER)) {
+            return new ArrayList<>();
+        }
+
         ArrayList<WebsitePromotionVO> voarr = new ArrayList<>();
         try {
             ArrayList<WebsitePromotionPO> poarr = websitePromotionDataService.getAll();
@@ -37,6 +48,15 @@ public class WebsitePromotionBL implements WebsitePromotionBLService {
 
     @Override
     public ResultMessage add(WebsitePromotionVO vo) {
+        ResultMessage checkResult = checkVO(vo);
+        if (checkResult.result != ResultMessage.RESULT_SUCCESS) {
+            return checkResult;
+        }
+
+        if (session != null && !session.userType.equals(UserType.WEBSITE_MARKETER)) {
+            return new ResultMessage(ResultMessage.RESULT_ACCESS_DENIED);
+        }
+
         WebsitePromotionPO po = WebsitePromotionBL.toPO(vo);
         try {
             websitePromotionDataService.insert(po);
@@ -49,6 +69,10 @@ public class WebsitePromotionBL implements WebsitePromotionBLService {
 
     @Override
     public ResultMessage delete(int promotionId) {
+        if (session != null && !session.userType.equals(UserType.WEBSITE_MARKETER)) {
+            return new ResultMessage(ResultMessage.RESULT_ACCESS_DENIED);
+        }
+
         try {
             WebsitePromotionPO po = websitePromotionDataService.getById(promotionId);
             websitePromotionDataService.delete(po);
@@ -61,6 +85,15 @@ public class WebsitePromotionBL implements WebsitePromotionBLService {
 
     @Override
     public ResultMessage modify(WebsitePromotionVO vo) {
+        ResultMessage checkResult = checkVO(vo);
+        if (checkResult.result != ResultMessage.RESULT_SUCCESS) {
+            return checkResult;
+        }
+
+        if (session != null && !session.userType.equals(UserType.WEBSITE_MARKETER)) {
+            return new ResultMessage(ResultMessage.RESULT_ACCESS_DENIED);
+        }
+
         WebsitePromotionPO po = WebsitePromotionBL.toPO(vo);
         try {
             websitePromotionDataService.update(po);
@@ -69,6 +102,19 @@ public class WebsitePromotionBL implements WebsitePromotionBLService {
             // e.printStackTrace();
             return new ResultMessage(ResultMessage.RESULT_DB_ERROR);
         }
+    }
+
+    private ResultMessage checkVO(WebsitePromotionVO vo) {
+        if (vo.name.trim().isEmpty()) {
+            return new ResultMessage("促销策略名称不能为空，请重新输入");
+        }
+        if (vo.promotion < 0 || vo.promotion > 1) {
+            return new ResultMessage("折扣幅度应为 0~1 之间的小数，请重新输入");
+        }
+        if (vo.startTime.compareTo(vo.endTime) > 0) {
+            return new ResultMessage("促销开始时间应不晚于结束时间，请重新输入");
+        }
+        return new ResultMessage(ResultMessage.RESULT_SUCCESS);
     }
 
     private static WebsitePromotionVO toVO(WebsitePromotionPO po) {
